임시저장
몬스터 기본 기능 구현 완료
   - 플레이어 탐지 및 추적
   - Pool로 소환
기본 무기 기능 일부 구현 완료
   - 무기 생성
   - 근접 몬스터 탐지
TODO:
   - 발사체 구현하기
   - 피격 구현하기

+++ My_Profile/Assets/Scripts/Core/Utills/SO/BaseData.cs
@@ -4,6 +4,6 @@ public  abstract class BaseData : ScriptableObject
     [Header("Character Data")]
     public string Name;
     public string Description;
-    public int MaxHP;
+    public float MaxHP;
     public float MoveSpeed;
 }
